feat(client): suspense 적용
